refactor: extract user sync logic
